codegen-poc/requirements.txt
openai==1.65.2
python-dotenv

